+++ NEWS.md
@@ -87,6 +87,8 @@
   `$load()`, `$update()`, `$paste()` methods in `Idf` class and `$set()` method
   in `IdfObject` class (#133).
 * EnergyPlus v9.2 support has been added (#138).
+* Now required objects and unique objects can be deleted in `$del()` when
+  `.force` is set to `TRUE` (#149).
 
 ## Bug fixes